Sort out how to name properties in command responses.
Every variable (non-static) will now always come back with a $ prefixed,
regardless of whether the $ was part of the -n argument to property_get.

Every static property in the current scope can be requested by
"::propertyName". "::" can not be requested, but "$t::" can in which case it
returns only all static properties of the class. Private class properties will
be named "::*parentClass*propertyName", and can also be requested with
property_get.

I've also adjusted all the test cases that checked for the old behaviour.