fix endless recursive calls of xdebug_error_cb
Reproducible with SOAP ext and (not sure if that matters) PHP 5.5.
The first execution of this code used to set zend_error_cb to
xdebug_error_cb.
The second execution of it set also xdebug_external_error_cb to
xdebug_error_cb, which resulted in xdebug_error_cb calling itself
till PHP crashed.