Cater to PHP patchlevel releases in AppVeyor CI
Windows patchlevel releases (i.e. releases which only update dependency
libs) have a trailing `-<num>` in the version.  However, the toplevel
folder of the devel packages does not have that trailing `-<num>`.
We cater to that by renaming that folder after unpacking to work
smoothly with the rest of the CI scripts.