Install zlib for GH CI on Windows
A suitable zlib package needs to be available for zlib support on
Windows (PR #782).  This patch downloads and unzips this to the default
dependency folder, so that it should be recognized and used without
further ado.

Note that this solution hard-codes the zlib version, which is no
problem for now, but might be in the future.  A clean solution would
need to download and parse the respective series file first, to find
out which zlib version is actually available.  However, that should
better be provided by cmb69/setup-php-sdk.