Quick-fix for missing tracing initialization on ZTS Windows
If tracing is done in a ZTS environment on Windows 8.0+,
xdebug_nanotime_context.win_precise_time_func is not properly
initialized, and the wrong code path is taken, resulting in a division
by zero, which causes the Apache worker process to be terminated.

This is a quick-fix to solve this issue, which may at least serve to
figure out whether bugs 2016, 2924 and 2028 actually have this very
root cause.  A proper fix might be to make xdebug_nanotime_context a
proper module global, and to initialize it in GINIT.