Fixed issue #2027: Branch/path code coverage for traits misses trait name
When analysing code, the scope is the *trait* name. While the code is running,
the method has been copied into the class, and the scope is the *class* name.

I cheated earlier and removed the scope name to make code coverage work.

This "fix" is more of a plaster. We now remember the original scope and then
add it when returning information.

PHP doesn't remember the original trait scope *yet*, but we might fix that
as part of this PR: https://github.com/php/php-src/pull/12129