Backport: Skip bug02020.phpt on Windows
This failure is caused by the fix for GHSA-p99j-rfp4-xqvq.  A query
string not containing an `=` causes the argument parsing to be
disabled, so the `--INI--` section does not have any effect on Windows.
Since there is no way to workaround that, we skip the test on Windows
unconditionally.